Finished changed on the global project structure
