--- SDSPy/Screen/__init__.py
@@ -1,1 +1,0 @@
-from .screen import SiglentScreen
