+++ SDSPy/WGEN/__init__.py
@@ -1,0 +1,1 @@
+from .wgen import SiglentWGEN
