--- SDSPy/counter/__init__.py
@@ -1,1 +1,0 @@
-from .counter import SiglentCounter
