+++ SDSPy/decode/__init__.py
@@ -1,0 +1,1 @@
+from .decode import SiglentDecode
