+++ SDSPy/deprecated_functions/__init__.py
@@ -1,2 +1,3 @@
 from .acal import ACAL
 from .date import DATE
+from .counter import COUNTER
