+++ SDSPy/deprecated_functions/counter.py
@@ -1,0 +1,13 @@
+# ============================================================================================================
+# Counter.py
+# lheywang on 19/12/2024
+#
+# Base file for the counter class
+#
+# ============================================================================================================
+
+
+class COUNTER:
+    def __init__(self, instr, baseclass):
+        self.__instr__ = instr
+        self.__baseclass__ = baseclass
