+++ SDSPy/digital/__init__.py
@@ -1,0 +1,1 @@
+from .digital import SiglentDigital
