+++ SDSPy/digital/digital.py
@@ -1,0 +1,12 @@
+# ============================================================================================================
+# Digital.py
+# lheywang on 21/12/2024
+#
+# Base file for the digital class
+#
+# ============================================================================================================
+
+class SiglentDigital:
+    def __init__(self, instr, baseclass):
+        self.__instr__ = instr
+        self.__baseclass__ = baseclass
