+++ SDSPy/generics/__init__.py
@@ -1,0 +1,1 @@
+from .generics import SCPIGenerics
