+++ SDSPy/measure/measure.py
@@ -1,13 +1,12 @@
 # ============================================================================================================
-# Counter.py
-# lheywang on 19/12/2024
+# Measure.py
+# lheywang on 21/12/2024
 #
-# Base file for the counter class
+# Base file for the measure class
 #
 # ============================================================================================================
 
-
-class SiglentCounter:
+class SiglentMeasure:
     def __init__(self, instr, baseclass):
         self.__instr__ = instr
         self.__baseclass__ = baseclass
