+++ SDSPy/waveform/__init__.py
@@ -1,0 +1,1 @@
+from .waveform import SiglentWaveform
